polygon refactor
